Fix

# Conflicts:
#	packages/language/src/language-server/semantic-tokens.ts
#	packages/language/src/parser/tokenizer/pli-tokenizer.ts
