+++ packages/language/src/parser/tokenizer.ts
@@ -23,6 +23,7 @@ import {
   tokenizeIncludeAlt,
   updatePliTokenizer,
 } from "./tokenizer/pli-tokenizer";
+import { updateCicsTokenizer } from "./tokenizer/cics-tokenizer";
 
 export interface TokenizationResult {
   tokens: tokens.Token[];
@@ -94,3 +95,4 @@ export function tokenize(
 }
 
 updatePliTokenizer(getDefaultCompilerOptions());
+updateCicsTokenizer();
