+++ packages/language/src/preprocessor/pli-lexer.ts
@@ -31,6 +31,7 @@ import { getDefaultCompilerOptions } from "./compiler-options/options";
 import { CompilerOptions } from "./compiler-options/options-pli";
 import { DiagnosticCategory } from "../validation/diagnostics-store";
 import { updatePliTokenizer } from "../parser/tokenizer/pli-tokenizer";
+import { updateCicsTokenizer } from "../parser/tokenizer/cics-tokenizer";
 
 export interface LexerResult {
   all: Token[];
@@ -65,6 +66,7 @@ export class PliLexer {
       compilerOptionsResult.result?.options ?? getDefaultCompilerOptions();
     unit.compilerOptions = opts;
     updatePliTokenizer(opts);
+    updateCicsTokenizer();
     unit.instructionCache.update(opts);
     const instruction = unit.instructionCache.get(uri, inputText, () => {
       const textWithoutMargins = this.marginsProcessor.processMargins(
